+++ config/presets.mjs
@@ -116,6 +116,10 @@ presets.MORIO_CONTAINER_TAG = 'dev-build'
 predocs.MORIO_RELEASE_CHANNEL = 'The current release channel. Picked up from environment variables.'
 presets.MORIO_RELEASE_CHANNEL = 'dev'
 
+predocs.MORIO_TEMPLATE_TAGS = 'Custom Mustache delmiters to use'
+presets.MORIO_TEMPLATE_TAGS = ['{|', '|}']
+
+
 /*
  * Docker presets
  */
@@ -289,20 +293,8 @@ presets.MORIO_CACHE_PORT = 6379
  * Connnector presets
  */
 
-predocs.MORIO_CONNECTOR_LOGSTASH_PIPELINE_FOLDER =
-  'The location of the pipeline configurations within the Logstash container.'
-presets.MORIO_CONNECTOR_LOGSTAHS_PIPELINE_FOLDER = '/usr/share/logstash/config/pipeline'
-
-predocs.MORIO_CONNECTOR_LOGSTASH_PIPELINE_CONFIG_FILE =
-  'The location of the central pipeline configuration file within the Logstash container.'
-presets.MORIO_CONNECTOR_LOGSTASH_PIPELINE_CONFIG_FILE = '/etc/morio/connector/pipelines.yml'
-
 predocs.MORIO_CONNECTOR_UID = 'User ID (UID) of the user running the Connector service'
 presets.MORIO_CONNECTOR_UID = 1000
-
-predocs.MORIO_CONNECTOR_P12_PASSWORD =
-  'A password to protect PKCS12 keystores. Does not provide actual security.'
-presets.MORIO_CONNECTOR_P12_PASSWORD = 'Butterflies and zebras and moonbeams'
 
 /*
  * Console presets