+++ core/src/lib/tls.mjs
@@ -178,8 +178,10 @@ export async function ensureServiceCertificate(service, internal = false, chain 
    */
   if (json && missing < 1) {
     const days = Math.floor((new Date(json.expires).getTime() - Date.now()) / (1000 * 3600 * 24))
-    if (days > 66) return true
-    else log.info(`[${service}] TLS certificate will expire in ${days}. Renewing now.`)
+    if (days > 66) {
+      log.debug(`[${service} TLS certificate will expire in ${days} days. No renewing.`)
+      return true
+    } else log.info(`[${service}] TLS certificate will expire in ${days}. Renewing now.`)
   }
 
   /*
@@ -233,12 +235,14 @@ export async function ensureServiceCertificate(service, internal = false, chain 
   /*
    * Also write broker certificates to the downloads folder
    */
-  await writeFile(`/morio/data/downloads/certs/${service}.pem`, certAndKey.certificate.crt)
+  if (service === 'broker') {
+    await writeFile(`/morio/data/downloads/certs/${service}.pem`, certAndKey.certificate.crt)
+  }
 
   /*
    * And finally, write a JSON file to keep track of certificate expiry
    */
-  await writeJsonFile('/etc/morio/db/certs.json', {
+  await writeJsonFile(`/etc/morio/${service}/certs.json`, {
     created: new Date(),
     expires: new Date(
       Date.now() + certificateLifetimeInMs(utils.getPreset('MORIO_CA_CERTIFICATE_LIFETIME_MAX'))