+++ shared/src/loaders.mjs
@@ -8,7 +8,6 @@ import { rm, mkdir, readFile, writeFile, globDir } from './fs.mjs'
 import { cloneAsPojo, get, set, setIfUnset, reverseString } from './utils.mjs'
 import merge from 'lodash/merge.js'
 import unset from 'lodash/unset.js'
-import mustache from 'mustache'
 
 /*
  * A collection of utils to load various files
@@ -806,12 +805,3 @@ function applyOverlay(settings, overlay = {}) {
   return settings
 }
 
-export function extractTemplateVariables(template) {
-  const vars = []
-  for (const chunk of mustache.parse(template)) {
-    if (chunk[0] === '&' && !chunk[1].includes(' ')) vars.push(chunk[1])
-    else if (chunk[0] === 'name') vars.push(chunk[1])
-  }
-
-  return vars
-}