+++ ui/hooks/use-template.mjs
@@ -2,7 +2,9 @@ import mustache from 'mustache'
 
 export const useTemplate = (data) => {
   const template = (input, replace) =>
-    typeof input === 'undefined' ? input : mustache.render(input, { ...data, ...replace })
+    typeof input === 'undefined'
+      ? input
+      : mustache.render(input, { ...data, ...replace }, { tags: ['{|', '|}'] })
 
   return template
 }